implemented setup file
